+++ presentation/src/main/java/com/nstu/technician/feature/ContainerActivity.kt
@@ -21,7 +21,7 @@ class ContainerActivity : BaseActivity() {
 
         private const val EXTRA_TECHNICIAN_ID = "technicianId"
 
-        fun startActivity(activity: BaseActivity, userId: Int) {
+        fun startActivity(activity: BaseActivity, userId: Long) {
             val intent = Intent(activity, ContainerActivity::class.java)
             intent.putExtra(EXTRA_TECHNICIAN_ID, userId)
             activity.startActivity(intent)
